+++ demos/dev/src/entity/Foreground.ts
@@ -64,12 +64,8 @@ export const foregroundArchetype: Archetype = {
             tilemap: ASSETS.tilemap.main,
             objects: tiledObjects,
         }),
+        // the tilesets are created by the TiledWrapper from the ones embedded in the tilemap
         new TilemapRenderer({
-            tileset: {
-                image: ASSETS.images.tileset,
-                tileWidth: 16,
-                tileHeight: 16,
-            },
             layer: RENDER_LAYERS.Foreground,
         }),
         new TilemapCollider({